Rename ophan.ts so that we can test functions in it
