+++ support-frontend/assets/components/paymentRequestButton/hooks/paymentAuthorisation.ts
@@ -11,11 +11,11 @@ import type { CsrfState } from 'helpers/redux/checkout/csrf/state';
 import { setPaymentRequestError } from 'helpers/redux/checkout/payment/paymentRequestButton/actions';
 import type { ContributionsDispatch } from 'helpers/redux/contributionsStore';
 import { trackComponentLoad } from 'helpers/tracking/behaviour';
-import { trackComponentEvents } from 'helpers/tracking/ophan';
 import {
 	onThirdPartyPaymentAuthorised,
 	paymentWaiting,
 } from 'pages/supporter-plus-landing/setup/legacyActionCreators';
+import { trackComponentEvents } from '../../../helpers/tracking/trackingOphan';
 
 export async function fetchClientSecret(
 	stripePublicKey: string,
