+++ support-frontend/assets/pages/[countryGroupId]/events/events.tsx
@@ -15,9 +15,9 @@ import { Header } from 'components/headers/simpleHeader/simpleHeader';
 import { PageScaffold } from 'components/page/pageScaffold';
 import { guardianLiveTermsLink, privacyLink } from 'helpers/legal';
 import * as cookie from 'helpers/storage/cookie';
-import { getPageViewId } from 'helpers/tracking/ophan';
 import { isProd } from 'helpers/urls/url';
 import type { GeoId } from 'pages/geoIdConfig';
+import { getPageViewId } from '../../../helpers/tracking/trackingOphan';
 
 const darkBackgroundContainerMobile = css`
 	background-color: ${palette.neutral[97]};
