+++ support-frontend/assets/pages/supporter-plus-landing/components/contributionsPriceCards.tsx
@@ -12,7 +12,7 @@ import {
 	useContributionsDispatch,
 	useContributionsSelector,
 } from 'helpers/redux/storeHooks';
-import { navigateWithPageView } from 'helpers/tracking/ophan';
+import { navigateWithPageView } from '../../../helpers/tracking/trackingOphan';
 
 const titleAndButtonContainer = css`
 	display: flex;
